+++ src/OFS/PropertySheets.py
@@ -470,7 +470,7 @@ class PropertySheets(Traversable, Implicit, Tabs):
     def get(self, name, default=None):
         for propset in self.__propsets__():
             if propset.id == name or \
-               getattr(propset, 'xml_namespace', None) == name:
+               getattr(propset, 'xml_namespace', object)() == name:
                 return propset.__of__(self)
         return default